translator/requirements.txt
openai
python-dotenv
bs4
lxml
python-dotenv